Move Handle and Now / SystemNow into timer mod
